fix: caption in Figure supports React node
