+++ packages/epo-react-lib/src/atomic/Figure/Figure.stories.tsx
@@ -3,6 +3,7 @@ import { children } from "@/storybook/utilities/argTypes";
 
 import Figure from ".";
 import Image from "../Image";
+import { generateCantoSrcSet } from "@/storybook/utilities/helpers";
 
 const meta: Meta<typeof Figure> = {
   component: Figure,
@@ -32,11 +33,15 @@ export const VerticalLayout: StoryObj<typeof Figure> = {
     children: (
       <Image
         image={{
-          url: "https://rubin.canto.com/direct/image/2378fnrgod2nnfha8mt8fool0q/Jn8gNFKti6C9EsMjscGaiHzWQlE/m800/800",
+          srcSet: generateCantoSrcSet(
+            "https://rubin.canto.com/direct/image/2378fnrgod2nnfha8mt8fool0q/Jn8gNFKti6C9EsMjscGaiHzWQlE/m800/",
+            4928
+          ),
+          url: "https://rubin.canto.com/direct/image/2378fnrgod2nnfha8mt8fool0q/j6VBew_jsPseN6cHtd4SD04xccU/original?content-type=image%2Fjpeg&name=Rubin_May_2022.jpg",
           altText:
             "Wide view of the Rubin Observatory facility and smaller surrounding buildings against an orange and purple twilight sky",
-          width: 800,
-          height: 530,
+          width: 4928,
+          height: 3264,
         }}
       />
     ),
