--- packages/epo-react-lib/src/atomic/Figure/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Figure";
